requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Pre-trained Models
transformers>=4.30.0
tokenizers>=0.13.0

# Web UI
gradio>=4.0.0

# NLP Libraries
gensim>=4.3.0
nltk>=3.8.0
spacy>=3.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Development
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# Utilities
tqdm>=4.65.0
rich>=13.3.0
